Now performs an rDNS lookup against the machine we're connecting to.
Previously, if a hostname had multiple IP addresses, there was a not
insignificant chance that the Kerberos library would do an rDNS lookup
and find a hostname that didn't match what we were connecting to.

We now do our own rDNS lookup against our socket peer.

This is modelled on https://github.com/cannatag/ldap3/pull/89, which
fixes support for DNS round-robins in the ldap3 library.